refactor project structure
mappers belong to application
